+++ src/common/components/FieldIsPrimary/FieldIsPrimary.js
@@ -1,19 +1,18 @@
 import React from 'react';
-import { Field } from 'redux-form';
+import { Field, useForm } from 'react-final-form';
 import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
 
 import { Checkbox } from '@folio/stripes/components';
 
-function FieldIsPrimary({ dispatchChange, fields, fieldIndex, fieldPrefix, labelId, vertical }) {
-  const changeIsPrimary = (e, newValue) => {
-    if (newValue) {
-      fields.forEach((fieldName, i) => {
-        if (i !== fieldIndex && fields.get(i).isPrimary === newValue) {
-          dispatchChange(`${fieldName}.isPrimary`, !newValue);
-        }
-      });
-    }
+function FieldIsPrimary({ fields, fieldIndex, fieldPrefix, labelId, vertical }) {
+  const { change } = useForm();
+  const changeIsPrimary = ({ target: { checked } }) => {
+    fields.forEach((fieldName, i) => (
+      i === fieldIndex
+        ? change(`${fieldName}.isPrimary`, checked)
+        : change(`${fieldName}.isPrimary`, false)
+    ));
   };
 
   return (
@@ -31,7 +30,6 @@ function FieldIsPrimary({ dispatchChange, fields, fieldIndex, fieldPrefix, label
 }
 
 FieldIsPrimary.propTypes = {
-  dispatchChange: PropTypes.func.isRequired,
   fieldIndex: PropTypes.number.isRequired,
   fieldPrefix: PropTypes.string.isRequired,
   fields: PropTypes.object.isRequired,